[update] split test code
